+++ packages/Fable.Form.Simple.Bulma/CHANGELOG.md
@@ -14,4 +14,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Fable.Form.Simple.Feliz.Bulma
+* Fable.Form.Simple.Bulma